Safari-609.1.11.1/LayoutTests/platform/gtk/accessibility/aria-labelledby-overrides-label-expected.txt

This tests that if aria-labelledby is used, then label elements are not used

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS platformValueForW3CName(text) is "Shut down computer after 10 minutes"
PASS text.titleUIElement() != null && text.titleUIElement().isValid is false
Label element role is: AXRole: AXLabel
PASS successfullyParsed is true

TEST COMPLETE
